pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "editorbot"
version = "0.1.0"
description = "A Telegram bot for text mediation with dialect conversion"
requires-python = ">=3.11,<3.13"
authors = [
    {name = "Editor Bot Team"}
]
license = {text = "MIT"}
dependencies = [
    "python-telegram-bot>=21.0",
    "pydub>=0.25.1",
    "google-generativeai>=0.6.0",
    "google-cloud-speech>=2.30.0",
    "ffmpeg-python>=0.2.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0",
    "ruff>=0.1.0"
]
local-transcription = [
    "openai-whisper>=20240930"
]

[tool.setuptools.packages.find]
where = ["."]

[project.scripts]
editorbot = "bot.bot:main"

